Add responsive classes to blog article img & iframes
Images and embedded video were not displaying in blog articles using
Bootstrap responsive classes.

I was annoyed by having to add these classes to blog posts by hand, and
decided to add this functionality in the template file.

In this update, the embed-responsive-16by9 class is applied to all
iframe elements regardless of actual ratio.

https://getbootstrap.com/docs/4.0/utilities/embed/