jquery-zoom shows full-size product image. Close #10
The jquery-zoom plugin was not displaying the true full-size product
image on mouseover. Instead it was attaching to an image displayed by
the responsive-image snippet. This resulted in some zoom, but not the
greatest amount possible.

In this update to product.js, the full-size image URL is extracted from
the active thumbnail element underneath the main product image, and
passed to jquery-zoom.

The JS regex assumes these thumbnails have _300x appended to the src
element.