Add data-product-form attributes to featured-product
Previously, data-section-type="product" was removed from the
featured-production section, because it was loading sections/product.js
on the home page, resulting in errors.

This update restores data-section-type, and adds data-product-handle. No
errors occur on page load with product.js and products are successfully
added to cart.

Future update may attempt ajax update of cart when item is added.